Changed the swagger ui and api endpoint files so everything wasn't in the same file Changed instructional text to support changes to swagger ui Removed non-existent swagger endpoint Changed /health to a variable instead of using the string everywhere Reduced the complexity of the doFilterInternal method in JwtTokenAuthenticationFilter
