+++ api/src/main/java/gov/cms/ab2d/api/controller/HealthAPI.java
@@ -7,12 +7,14 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import static gov.cms.ab2d.common.util.Constants.HEALTH_ENDPOINT;
+
 @RestController
 public class HealthAPI {
     @Autowired
     private HealthCheck healthCheck;
 
-    @GetMapping("/health")
+    @GetMapping(HEALTH_ENDPOINT)
     public ResponseEntity<Void> getHealth() {
         if (healthCheck.healthy()) {
             return new ResponseEntity<>(HttpStatus.OK);
