+++ api/src/main/java/gov/cms/ab2d/api/security/SecurityConfig.java
@@ -50,7 +50,7 @@ public class SecurityConfig extends WebSecurityConfigurerAdapter {
     @Override
     public void configure(WebSecurity web) {
         web.ignoring().antMatchers("/swagger-ui.html", "/swagger-ui.html/**", "/configuration/**",
-                "/swagger-resources/**", "/v2/api-docs", "/webjars/**", "/health", STATUS_ENDPOINT);
+                "/swagger-resources/**", "/v2/api-docs", "/webjars/**", HEALTH_ENDPOINT, STATUS_ENDPOINT);
     }
 
     @Override
